+++ src/test/java/dinkplugin/notifiers/LootNotifierTest.java
@@ -653,6 +653,69 @@ class LootNotifierTest extends MockedNotifierTest {
     }
 
     @Test
+    void testNotifyGauntlet() {
+        // prepare data
+        int kc = 123;
+        int quantity = 24;
+        String total = QuantityFormatter.quantityToStackSize(quantity * RUBY_PRICE);
+        String source = "The Gauntlet";
+        List<ItemStack> items = List.of(new ItemStack(ItemID.RUBY, quantity, null));
+
+        // fire events
+        killCountService.onGameMessage(String.format("Your Gauntlet completion count is: %d.", kc));
+        plugin.onLootReceived(new LootReceived(source, -1, LootRecordType.EVENT, items, 1));
+
+        // verify notification message
+        verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
+            false,
+            NotificationBody.builder()
+                .text(
+                    Template.builder()
+                        .template(String.format("%s has looted: %d x {{ruby}} (%s) from {{source}} for %s gp", PLAYER_NAME, quantity, total, total))
+                        .replacement("{{ruby}}", Replacements.ofWiki("Ruby"))
+                        .replacement("{{source}}", Replacements.ofWiki(source))
+                        .build()
+                )
+                .extra(new LootNotificationData(List.of(new SerializedItemStack(ItemID.RUBY, quantity, RUBY_PRICE, "Ruby")), source, LootRecordType.EVENT, kc, null))
+                .type(NotificationType.LOOT)
+                .build()
+        );
+    }
+
+    @Test
+    void testNotifyCorruptedGauntlet() {
+        // prepare data
+        int kc = 123;
+        int quantity = 24;
+        String total = QuantityFormatter.quantityToStackSize(quantity * RUBY_PRICE);
+        String source = "The Gauntlet";
+        String realSource = "Corrupted Gauntlet";
+        List<ItemStack> items = List.of(new ItemStack(ItemID.RUBY, quantity, null));
+
+        // fire events
+        killCountService.onGameMessage(String.format("Your Corrupted Gauntlet completion count is: %d.", kc));
+        plugin.onLootReceived(new LootReceived(source, -1, LootRecordType.EVENT, items, 1));
+
+        // verify notification message
+        verify(messageHandler).createMessage(
+            PRIMARY_WEBHOOK_URL,
+            false,
+            NotificationBody.builder()
+                .text(
+                    Template.builder()
+                        .template(String.format("%s has looted: %d x {{ruby}} (%s) from {{source}} for %s gp", PLAYER_NAME, quantity, total, total))
+                        .replacement("{{ruby}}", Replacements.ofWiki("Ruby"))
+                        .replacement("{{source}}", Replacements.ofWiki(realSource))
+                        .build()
+                )
+                .extra(new LootNotificationData(List.of(new SerializedItemStack(ItemID.RUBY, quantity, RUBY_PRICE, "Ruby")), realSource, LootRecordType.EVENT, kc, null))
+                .type(NotificationType.LOOT)
+                .build()
+        );
+    }
+
+    @Test
     void testDisabled() {
         // disable notifier
         when(config.notifyLoot()).thenReturn(false);